added authorization utility to routes

+++ controllers/api/userRoutes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const { User, Post } = require('../../models');
+const withAuth = require('../../utils/auth');
 
 router.get('/', async (req, res) => {
   try {
@@ -99,8 +100,14 @@ router.post('/logout', (req, res) => {
   }
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', withAuth, async (req, res) => {
   try {
+
+    if (req.params.id != req.session.user_id) {
+      res.status(403).json({message: "You cannot modify other user's account"})
+      return;
+    }
+
     const userData = await User.update(req.body, {
       individualHooks: true,
       where: {
@@ -120,8 +127,14 @@ router.put('/:id', async (req, res) => {
   }
 });
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', withAuth, async (req, res) => {
   try {
+
+    if (req.params.id != req.session.user_id) {
+      res.status(403).json({message: "You cannot modify other user's account"})
+      return;
+    }
+    
     const userData = await User.destroy({
       where: {
         id: req.params.id
